agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.4.4",
    "livekit-plugins-cartesia>=1.4.4",
    "livekit-plugins-deepgram>=1.4.4",
    "livekit-plugins-elevenlabs>=1.4.4",
    "livekit-plugins-openai>=1.4.4",
    "python-dotenv>=1.2.2",
]
